requirement.txt
alembic==1.8.1
aniso8601==9.0.1
APScheduler==3.10.4
arrow==1.2.3
beautifulsoup4==4.11.2
better-profanity==0.7.0
blinker==1.9.0
boto3==1.28.40
botocore==1.31.40
CacheControl==0.14.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
DateTime==4.7
decorator==4.4.2
distlib==0.3.6
encode==0.0.0
extension==1.0.0
filelock==3.8.0
firebase-admin==6.5.0
Flask==2.2.5
Flask-APScheduler==1.13.1
flask-blueprint==1.3.0
Flask-CKEditor==0.4.6
Flask-Cors==3.0.10
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.5
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
geographiclib==2.0
geopy==2.3.0
google==3.0.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==2.0.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.8.1
h2==3.2.0
haversine==2.7.0
hpack==3.0.0
httplib2==0.22.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.10.0
moviepy==1.0.3
msgpack==1.1.0
mysqlclient==2.1.1
numpy==1.24.2
packaging==21.3
pandas==1.5.3
Pillow==9.3.0
platformdirs==2.5.4
profanity-check==1.0.3
proglog==0.1.10
proto-plus==1.26.0
protobuf==5.29.3
py-dotenv==0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.4
qrcode==8.2
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
sql==2022.4.0
SQLAlchemy==1.4.41
threadpoolctl==3.1.0
timeago==1.0.16
times==0.7
tqdm==4.66.4
types-cryptography==3.3.23.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.7
Werkzeug==2.3.8
WTForms==3.0.1
zope.interface==5.5.2
